refactor(voip): extract call forwarding orchestrator
